+++ tests/Mockery/DemeterChainTest.php
@@ -19,6 +19,10 @@
  * @license    http://github.com/padraic/mockery/blob/master/LICENSE New BSD License
  */
 
+if (version_compare(PHP_VERSION, '7.0.0') >= 0) {
+    require_once __DIR__.'/DummyClasses/DemeterChain.php';
+}
+
 use Mockery\Adapter\Phpunit\MockeryTestCase;
 
 class DemeterChainTest extends MockeryTestCase
@@ -182,5 +186,19 @@ class DemeterChainTest extends MockeryTestCase
         $this->assertEquals(2, mock1::select()->some()->other()->data());
         $this->assertEquals(3, mock2::select()->some()->data());
         $this->assertEquals(4, mock2::select()->some()->other()->data());
+    }
+
+    /**
+     * @requires PHP 7.0.0
+     */
+    public function testDemeterChainsWithClassReturnTypeHints()
+    {
+        $a = \Mockery::mock(\DemeterChain\A::class);
+        $a->shouldReceive('foo->bar->baz')->andReturn(new stdClass);
+
+        $m = new \DemeterChain\Main();
+        $result = $m->callDemeter($a);
+
+        $this->assertInstanceOf(stdClass::class, $result);
     }
 }