+++ tests/Mockery/ExpectationTest.php
@@ -1080,6 +1080,23 @@ class ExpectationTest extends MockeryTestCase
         $this->mock->foo(1, 2, 3);
     }
 
+    public function testAndAnyOtherConstraintMatchesTheRestOfTheArguments()
+    {
+        $this->mock->shouldReceive('foo')->with(1, 2, Mockery::andAnyOthers())->twice();
+        $this->mock->foo(1, 2, 3, 4, 5);
+        $this->mock->foo(1, 'str', 3, 4);
+    }
+
+    /**
+     * @expectedException \Mockery\Exception
+     */
+    public function testAndAnyOtherConstraintDoesNotPreventMatchingOfRegularArguments()
+    {
+        $this->mock->shouldReceive('foo')->with(1, 2, Mockery::andAnyOthers());
+        $this->mock->foo(10, 2, 3, 4, 5);
+        Mockery::close();
+    }
+
     public function testArrayConstraintMatchesArgument()
     {
         $this->mock->shouldReceive('foo')->with(Mockery::type('array'))->once();
@@ -1978,6 +1995,14 @@ class ExpectationTest extends MockeryTestCase
         $mock->shouldReceive('foo->bar->baz')->andReturn('Spam!');
         $demeter = new Mockery_UseDemeter($mock);
         $this->assertSame('Spam!', $demeter->doitWithArgs());
+    }
+
+    public function testShouldNotReceiveCanBeAddedToCompositeExpectation()
+    {
+        $mock = mock('Foo');
+        $mock->shouldReceive('a')->once()->andReturn('Spam!')
+             ->shouldNotReceive('b');
+        $mock->a();
     }
 
     public function testPassthruEnsuresRealMethodCalledForReturnValues()