+++ tests/Mockery/NamedMockTest.php
@@ -22,6 +22,8 @@
 require_once __DIR__.'/DummyClasses/Namespaced.php';
 
 use Mockery\Adapter\Phpunit\MockeryTestCase;
+use test\Mockery\Stubs\Animal;
+use test\Mockery\Stubs\Habitat;
 
 class NamedMockTest extends MockeryTestCase
 {
@@ -64,5 +66,21 @@ class NamedMockTest extends MockeryTestCase
             array('water' => true)
         );
         $this->assertTrue($gardener->water($cactus));
+    }
+
+    /**
+     * @test
+     * @requires PHP 7.0.0
+     */
+    public function it_gracefully_handles_namespacing()
+    {
+        $animal = Mockery::namedMock(
+            uniqid(Animal::class, false),
+            Animal::class
+        );
+
+        $animal->shouldReceive("habitat")->andReturn(new Habitat());
+
+        $this->assertInstanceOf(Habitat::class, $animal->habitat());
     }
 }